再テスト後修正

+++ data_governance/library/main_menu/subflow_controller/utils.py
@@ -368,17 +368,21 @@ def backup_zipfile(abs_root: str, research_flow_dict: dict, current_time: str):
             notebook_list = get_notebook_list(working_path)
 
             with zipfile.ZipFile(zip_file_path, 'w', zipfile.ZIP_DEFLATED) as zipf:
-                for file in os.listdir(image_folder):
-                    file_path = os.path.join(image_folder, file)
-                    if os.path.isfile(file_path):
-                        zip_path = os.path.join(path_config.IMAGES, file)
-                        zipf.write(file_path, zip_path)
+                if notebook_list:
+                    for file in os.listdir(image_folder):
+                        file_path = os.path.join(image_folder, file)
+                        if os.path.isfile(file_path):
+                            zip_path = os.path.join(path_config.IMAGES, file)
+                            zipf.write(file_path, zip_path)
                 if os.path.exists(menu_notebook_path):
                     zipf.write(menu_notebook_path, arcname=os.path.basename(menu_notebook_path))
                 if os.path.exists(status_json_path):
                     zipf.write(status_json_path, arcname=os.path.basename(status_json_path))
                 for notebook in notebook_list:
                     zipf.write(notebook, arcname=os.path.basename(notebook))
+            with zipfile.ZipFile(zip_file_path, 'r') as zipf:
+                if not zipf.namelist:
+                    os.remove(zip_file_path)
 
 
 def get_govsheet_rf_path(abs_root: str) -> str:
@@ -635,7 +639,8 @@ def recreate_subflow(abs_root: str, govsheet_rf_path: str, govsheet_rf: dict, go
     for phase_name, subflow_data in research_flow_dict.items():
         for subflow_id, subflow_name in subflow_data.items():
             working_path = get_working_path(abs_root, phase_name, subflow_id)
-            shutil.rmtree(working_path)
+            if os.path.exists(working_path):
+                shutil.rmtree(working_path)
             for delete_file_name in delete_files:
                 delete_file_path = os.path.join(researchflow_path, phase_name, subflow_id, delete_file_name)
                 if not os.path.exists(delete_file_path):
